#140 Agregando analisis si es bien o servicio, separando recursos para validadores
